Fix XarrayStream getting passed an open Dataset, add a test
